AI-Media - all in one script for generating images, videos and audio from user prompts
Work in progress... still lots of work pending and lots of testing and fixing on Mac as the models work best on NVIDIA CUDA with known issues and false positives on Apple MPS (workarounds and fixes are being implemented so it also works on Mac). So far own testing and improvements on image generation only.